WebApp/requirements.txt
flask==2.3.3
flask-wtf==1.1.1
python-dotenv==1.0.1
requests==2.31.0
pillow==10.0.0
openai==1.3.0
azure-ai-formrecognizer==3.3.0
azure-storage-blob==12.17.0
gunicorn==21.2.0
python-multipart==0.0.6 